requirements.txt
Django==4.0.1
python-dotenv==0.19.2
psycopg2-binary==2.9.3
djangorestframework==3.13.1
social-auth-app-django==5.0.0
django-filter==2.4.0
pydantic==1.10.2
requests==2.27.1
pytest-django==4.5.2
pytest-factoryboy==2.5.0
